feat: aws setup for diff regions

+++ platform/apps/server/src/controlers/test-controllers.ts
@@ -18,7 +18,9 @@ export const deleteEc2ServerById = catchAsync(
     const { id }: { id?: string } = req.params;
     if (!id) throw new Error("id is requried");
 
-    const response = await terminateEc2Instance([id]);
+    const [instance] = await db.select().from(ec2).where(eq(ec2.ec2_id, id));
+    if (!instance) throw new Error("Instance is not found ");
+    const response = await terminateEc2Instance(instance.region, [id]);
     await db.delete(ec2).where(eq(ec2.ec2_id, id));
 
     res.status(200).json({
@@ -29,12 +31,12 @@ export const deleteEc2ServerById = catchAsync(
 
 export const createEc2Server = catchAsync(
   async (req: Request, res: Response) => {
-    const ec2res = await createEc2Instance({ region: "us-east-1" });
+    const ec2res = await createEc2Instance({ region: "ap-south-1" });
     for (const instance of ec2res.Instances ?? []) {
       if (instance.InstanceId) {
         await db.insert(ec2).values({
           ec2_id: instance.InstanceId,
-          region: "us-east-1",
+          region: "ap-south-1",
           ip: instance.PrivateIpAddress || null,
           status: "running",
         });
